+++ webui/src/lib/components/repository/compareBranchesActionBar.jsx
@@ -1,12 +1,8 @@
 import React, {useCallback, useRef, useState} from "react";
-import {refs, refs as refsAPI} from "../../../lib/api";
-import Alert from "react-bootstrap/Alert";
-import {RefTypeBranch, RefTypeCommit} from "../../../constants";
-import {useAPIWithPagination} from "../../hooks/api";
-import {appendMoreResults} from "../../../pages/repositories/repository/changes";
-import {ActionGroup, ActionsBar, AlertError, Loading, RefreshButton} from "../controls";
-import {ChangesTreeContainer, defaultGetMoreChanges, MetadataFields} from "./changes";
-import {URINavigator} from "./tree";
+import {refs as refsAPI} from "../../../lib/api";
+import {RefTypeBranch} from "../../../constants";
+import {ActionGroup, ActionsBar, AlertError, RefreshButton} from "../controls";
+import {MetadataFields} from "./changes";
 import {useRouter} from "../../hooks/router";
 import RefDropdown from "./refDropdown";
 import {ArrowLeftIcon, ArrowSwitchIcon, GitMergeIcon} from "@primer/octicons-react";
@@ -17,66 +13,7 @@ import Modal from "react-bootstrap/Modal";
 import Form from "react-bootstrap/Form";
 import {FormControl, FormHelperText, InputLabel, MenuItem, Select} from "@mui/material";
 
-const BranchComparison = (
-    {repo, reference, compareReference, showActionsBar, prefix = "", onSelectRef, onSelectCompare}
-) => {
-    const [internalRefresh, setInternalRefresh] = useState(true);
-
-    const [afterUpdated, setAfterUpdated] = useState(""); // state of pagination of the item's children
-    const [resultsState, setResultsState] = useState({prefix, results: [], pagination: {}}); // current retrieved children of the item
-
-    const delimiter = "/"
-
-    const {nextPage, loading, error} = useAPIWithPagination(async () => {
-        if (!repo) return
-
-        if (compareReference.id === reference.id) {
-            return {pagination: {has_more: false}, results: []}; // nothing to compare here.
-        }
-
-        const getMoreResults = () =>
-            refsAPI.diff(repo.id, reference.id, compareReference.id, afterUpdated, prefix, delimiter);
-        return await appendMoreResults(resultsState, prefix, afterUpdated, setAfterUpdated, setResultsState, getMoreResults);
-    }, [repo.id, reference.id, internalRefresh, afterUpdated, delimiter, prefix])
-
-    const {results} = resultsState;
-    const apiResult = {results, loading, error, nextPage};
-
-    const isEmptyDiff = (!loading && !error && !!results && results.length === 0);
-
-    const doRefresh = () => {
-        setResultsState({prefix, results: [], pagination: {}})
-        setInternalRefresh(!internalRefresh)
-    }
-
-    return (
-        <>
-            {showActionsBar &&
-                <CompareActionsBar
-                    repo={repo}
-                    reference={reference}
-                    onSelectRef={onSelectRef}
-                    compareReference={compareReference}
-                    onSelectCompare={onSelectCompare}
-                    doRefresh={doRefresh}
-                    isEmptyDiff={isEmptyDiff}
-                />
-            }
-            <BranchChangesList
-                apiResult={apiResult}
-                repo={repo}
-                reference={reference}
-                compareReference={compareReference}
-                prefix={prefix}
-                delimiter={delimiter}
-                refresh={internalRefresh}
-                setAfterUpdated={setAfterUpdated}
-            />
-        </>
-    );
-};
-
-const CompareActionsBar = (
+const CompareBranchesActionsBar = (
     {repo, reference, onSelectRef, compareReference, onSelectCompare, doRefresh, isEmptyDiff}
 ) => {
     const router = useRouter();
@@ -179,7 +116,7 @@ const MergeButton = ({repo, onDone, source, dest, disabled = false}) => {
         }
         setMergeState({merging: true, show: mergeState.show, err: mergeState.err, strategy: mergeState.strategy})
         try {
-            await refs.merge(repo.id, source, dest, strategy, message, metadata);
+            await refsAPI.merge(repo.id, source, dest, strategy, message, metadata);
             setMergeState({
                 merging: mergeState.merging,
                 show: mergeState.show,
@@ -243,97 +180,4 @@ const MergeButton = ({repo, onDone, source, dest, disabled = false}) => {
     );
 }
 
-const BranchChangesList = (
-    {apiResult, repo, reference, compareReference, prefix, delimiter, refresh, setAfterUpdated}
-) => {
-    const {results, loading, error, nextPage} = apiResult;
-
-    if (loading) return <Loading/>;
-    if (error) return <AlertError error={error}/>;
-
-    const changesTreeMessage =
-        <p>
-            Showing changes between <strong>{reference.id}</strong> {""}
-            and <strong>{compareReference.id}</strong>
-        </p>
-
-    if (compareReference.id === reference.id) {
-        return <Alert variant="warning">
-            <Alert.Heading>There isn’t anything to compare.</Alert.Heading>
-            You’ll need to use two different sources to get a valid comparison.
-        </Alert>;
-    }
-
-    return <ChangesTreeContainer
-        results={results}
-        reference={reference}
-        leftDiffRefID={getRefID(reference)}
-        rightDiffRefID={getRefID(compareReference)}
-        repo={repo}
-        delimiter={delimiter}
-        uriNavigator={CompareURINavigator(prefix, reference, compareReference, repo)}
-        internalReferesh={refresh}
-        prefix={prefix}
-        getMore={defaultGetMoreChanges(repo, reference.id, compareReference.id, delimiter)}
-        nextPage={nextPage}
-        setAfterUpdated={setAfterUpdated}
-        onNavigate={getNavigatorToComparePage(repo, reference, compareReference)}
-        changesTreeMessage={changesTreeMessage}
-    />;
-};
-
-function getURINavigatorRelativeTitle(from, to) {
-    let fromId = from.id;
-    let toId = to.id;
-    if (from.type === RefTypeCommit) {
-        fromId = fromId.substr(0, 12);
-    }
-    if (to.type === RefTypeCommit) {
-        toId = toId.substr(0, 12);
-    }
-    return `${fromId}...${toId}`
-}
-
-const CompareURINavigator = (prefix, reference, compareReference, repo) =>
-    <URINavigator
-        path={prefix}
-        reference={reference}
-        relativeTo={getURINavigatorRelativeTitle(reference, compareReference)}
-        repo={repo}
-        pathURLBuilder={(params, query) => {
-            const q = {
-                delimiter: "/",
-                prefix: query.path,
-            };
-            if (compareReference) {
-                q.compare = compareReference.id;
-            }
-            if (reference) {
-                q.ref = reference.id;
-            }
-            return {
-                pathname: '/repositories/:repoId/compare',
-                params: {repoId: repo.id},
-                query: q
-            };
-        }}/>;
-
-const getNavigatorToComparePage = (repo, ref, compareRef) => entry => ({
-    pathname: `/repositories/:repoId/compare`,
-    params: {repoId: repo.id},
-    query: {
-        ref: ref.id,
-        compare: compareRef.id,
-        prefix: entry.path,
-    }
-});
-
-function getRefID(reference) {
-    let refID = reference.id;
-    if (reference.type === RefTypeBranch) {
-        refID += "@";
-    }
-    return refID;
-}
-
-export default BranchComparison;
+export default CompareBranchesActionsBar;
